# Performance and memory usage fixes