--- a/lib/active_model_serializers/adapter.rb
+++ b/lib/active_model_serializers/adapter.rb
@@ -37,7 +37,7 @@ module ActiveModelSerializers
 
       # @return [Array<Symbol>] list of adapter names
       def adapters
-        adapter_map.keys.sort
+        adapter_map.keys.sort!
       end
 
       # Adds an adapter 'klass' with 'name' to the 'adapter_map'