@bf4 2019-01-29
rebase, add a note to the changelog, and we can merge